moor tidy of apps/amo
