--- apps/amo/templates/amo/breadcrumbs.html
@@ -1,11 +1,0 @@
-{% if breadcrumbs %}
-<ol id="breadcrumbs" class="breadcrumbs">
-  {% for target, label in breadcrumbs %}
-    {% if target %}
-      <li><a href="{{ target }}">{{ label }}</a></li>
-    {% else %}
-      <li>{{ label }}</li>
-    {% endif %}
-  {% endfor %}
-</ol>
-{% endif %}
