--- apps/amo/templates/amo/developer_webapp_agreement.html
@@ -1,59 +1,0 @@
-<!-- Not in a trans block on purpose; we don't want to get into
-   any legal ambiguities between the english version and a translation. -->
-
-<!-- When updating, don't forget to change the Last Updated date -->
-
-<div id="dev-agreement" class="rounded webapp">
-
-<strong>App Developer Preview Terms</strong>
-<p>Last Update: <strong>December 7, 2011</strong></p>
-
-<ol start="0">
-  <li>
-  <strong>The following terms</strong> (the &#8220;<strong>Terms</strong>&#8221;)
-  <strong>apply to your use of this developer preview of the Mozilla application
-    marketplace</strong> (the &#8220;<strong>App Developer Preview</strong>&#8221;).
-  By indicating your acceptance below or otherwise using the App Developer Preview,
-  you represent that you are at least 18 years of age, and have full power, capacity,
-  and authority to accept these Terms on behalf of yourself, or if applicable, your
-  employer or such other entity that you represent. If you do not agree to these Terms
-  or cannot make the foregoing representations, then you cannot use the App Developer
-  Preview. We reserve the right to change these Terms and the Firefox Market in our
-  sole discretion.
-  </li>
-  <li>
-  You may upload or otherwise make applications, add-ons or themes or other content
-  (&#8220;Developer Content&#8221;) available on the App Developer Preview in conjunction
-  with your testing of our service. In connection with the Developer Content you upload,
-  during the term of these Terms, you hereby grant us a nonexclusive, worldwide, royalty-free
-  license under all rights necessary to make such Developer Content available through the
-  functionality of the App Developer Preview.
-  </li>
-  <li>If any third party brings a claim against Mozilla related to your actions or your Developer
-  Content, you will defend and hold Mozilla harmless from and against all damages, losses, and
-  expenses of any kind (including reasonable legal fees and costs) related to such claim.</li>
-  <li>Mozilla reserves the right to manage the App Developer Preview in a manner designed to
-  facilitate its integrity and functionality, including the right to remove or disable Developer
-  Content or a developer&#39;s account.</li>
-  <li>Mozilla may terminate this Agreement or the App Developer Preview at any time. You may
-  terminate this Agreement by removing your Developer Content and ceasing use of the App Developer
-  Preview. Sections 2, and 4-7 will survive any termination or expiration of these Terms.</li>
-  <li class="loud-noises"><strong>The APP DEVELOPER PREVIEW IS provided &#8220;as is&#8221;.
-    Mozilla, its contributors, licensors and distributors disclaim all warranties,</strong>
-  whether express or implied, including without limitation, implied warranties of merchantability,
-  fitness for a particular purpose and non-infringement.  Some jurisdictions do not allow the exclusion
-  or limitation of implied warranties, so this disclaimer may not apply to you.</li>
-  <li class="loud-noises">YOU EXPRESSLY UNDERSTAND AND AGREE THAT, EXCEPT AS REQUIRED BY LAW, MOZILLA,
-  ITS CONTRIBUTORS, LICENSORS, SUBSIDIARIES, AFFILIATES AND DISTRIBUTORS WILL NOT BE LIABLE TO YOU UNDER
-  ANY THEORY OF LIABILITY FOR ANY INDIRECT, SPECIAL, INCIDENTAL, CONSEQUENTIAL, PUNITIVE OR EXEMPLARY DAMAGES
-  THAT MAY BE INCURRED BY YOU, INCLUDING ANY LOSS OF DATA, WHETHER OR NOT MOZILLA OR ITS REPRESENTATIVES HAVE
-  BEEN ADVISED OF OR SHOULD HAVE BEEN AWARE OF THE POSSIBILITY OF ANY SUCH LOSSES ARISING. IN NO EVENT SHALL
-  MOZILLA AGGREGATE LIABILITY TO YOU FOR ALL DAMAGES ARISING FROM OR RELATED TO THIS AGREEMENT BE IN EXCESS
-  OF $100.</li>
-  <li>This Agreement, and your relationship with Mozilla under this Agreement, shall be governed by the
-  laws of the State of California without regard to its conflict of laws provisions. You and Mozilla
-  agree to submit to the exclusive jurisdiction of the courts located within the county of Santa Clara,
-  California to resolve any legal matter arising from this Agreement.</li>
-</ol>
-
-</div>
