--- apps/amo/templates/services/install.html
@@ -1,44 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<title>Mozilla Add-ons</title>
-<script type="text/javascript">
-    function install() {
-        if ({{ referrers }}) {
-            var params = {{ params|safe }};
-            for (var key in params) params[key].toString = function () { return this.URL; };
-            InstallTrigger.install(params, goBack);
-        } else {
-            window.location.href = 'https://addons.mozilla.org/';
-        }
-    }
-
-    function goBack() {
-        history.back();
-    }
-</script>
-
-<style type="text/css">
-    .manualinstall {
-        text-align: center;
-        color: #b94413;
-        font-family:arial,verdana,sans-serif;
-        padding: 20px;
-        font-size: 17px;
-    }
-    .manualinstall a {
-        color: #b94413;
-        font-weight: bold;
-    }
-</style>
-</head>
-
-<body onload="install()">
-    <div class="manualinstall">
-      {% trans addon_name=addon['name'], addon_xpi=addon['xpi'] %}
-        If you are not prompted to install {{ addon_name }} in a moment, please <a href="{{ addon_xpi }}">click here</a>.
-      {% endtrans %}
-    </div>
-</body>
-</html>
-
