+++ apps/amo/tests/test_amo_utils.py
@@ -347,17 +347,3 @@ class TestAttachTransDict(amo.tests.TestCase):
             set([('en-us', 'English 2 Name'),
                  ('es', 'Spanish 2 Name'),
                  ('fr', 'French 2 Name')]))
-
-
-def test_has_links():
-    html = 'a text <strong>without</strong> links'
-    assert not amo.utils.has_links(html)
-
-    html = 'a <a href="http://example.com">link</a> with markup'
-    assert amo.utils.has_links(html)
-
-    html = 'a http://example.com text link'
-    assert amo.utils.has_links(html)
-
-    html = 'a badly markuped <a href="http://example.com">link'
-    assert amo.utils.has_links(html)
