+++ mkt/api/tests/test_authentication.py
@@ -8,13 +8,13 @@ from multidb.pinning import this_thread_is_pinned, unpin_this_thread
 from nose.tools import eq_, ok_
 from rest_framework.request import Request
 
-from amo.helpers import absolutify
 from amo.tests import TestCase
 from mkt.access.models import Group, GroupUser
 from mkt.api import authentication
 from mkt.api.middleware import RestOAuthMiddleware, RestSharedSecretMiddleware
 from mkt.api.models import Access, generate
 from mkt.api.tests.test_oauth import OAuthClient
+from mkt.site.helpers import absolutify
 from mkt.site.fixtures import fixture
 from mkt.site.middleware import RedirectPrefixedURIMiddleware
 from test_utils import RequestFactory
