+++ mkt/detail/helpers.py
@@ -2,8 +2,8 @@ import jinja2
 from jingo import env, register
 from tower import ungettext as ngettext
 
-from amo.helpers import numberfmt
 from mkt.constants.applications import DEVICE_TYPES
+from mkt.site.helpers import numberfmt
 
 
 @register.function
