+++ mkt/developers/tests/test_helpers.py
@@ -22,15 +22,15 @@ def test_hub_page_title():
 
     title = 'Oh hai!'
     s1 = render('{{ hub_page_title("%s") }}' % title, ctx)
-    s2 = render('{{ mkt_page_title("%s | Developers") }}' % title, ctx)
+    s2 = render('{{ page_title("%s | Developers") }}' % title, ctx)
     eq_(s1, s2)
 
     s1 = render('{{ hub_page_title() }}', ctx)
-    s2 = render('{{ mkt_page_title("Developers") }}', ctx)
+    s2 = render('{{ page_title("Developers") }}', ctx)
     eq_(s1, s2)
 
     s1 = render('{{ hub_page_title("%s", addon) }}' % title, ctx)
-    s2 = render('{{ mkt_page_title("%s | %s") }}' % (title, addon.name), ctx)
+    s2 = render('{{ page_title("%s | %s") }}' % (title, addon.name), ctx)
     eq_(s1, s2)
