+++ mkt/developers/tests/test_views_edit.py
@@ -19,7 +19,6 @@ from tower import strip_whitespace
 import amo
 import amo.tests
 import mkt
-from amo.helpers import absolutify
 from amo.tests import assert_required, formset, initial
 from amo.tests.test_helpers import get_image_path
 from lib.video.tests import files as video_files
@@ -29,6 +28,7 @@ from mkt.constants import regions
 from mkt.developers.models import ActivityLog
 from mkt.reviewers.models import RereviewQueue
 from mkt.site.fixtures import fixture
+from mkt.site.helpers import absolutify
 from mkt.translations.models import Translation
 from mkt.users.models import UserProfile
 from mkt.versions.models import Version
