+++ mkt/operators/helpers.py
@@ -3,7 +3,7 @@ from jingo import register
 
 from tower import ugettext_lazy as _lazy
 
-from mkt.developers.helpers import mkt_page_title
+from mkt.site.helpers import page_title
 
 
 @register.function
@@ -11,4 +11,4 @@ from mkt.developers.helpers import mkt_page_title
 def operators_page_title(context, title=None):
     section = _lazy('Operator Dashboard')
     title = u'%s | %s' % (title, section) if title else section
-    return mkt_page_title(context, title)
+    return page_title(context, title)
