+++ mkt/purchase/models.py
@@ -12,9 +12,11 @@ from tower import ugettext as _
 
 import amo
 from amo.fields import DecimalCharField
-from amo.helpers import absolutify, urlparams
+from amo.helpers import urlparams
 from amo.utils import get_locale_from_lang
+from mkt.site.helpers import absolutify
 from mkt.site.mail import send_mail, send_mail_jinja
+
 
 class ContributionError(Exception):
