+++ mkt/reviewers/utils.py
@@ -13,7 +13,6 @@ import waffle
 from tower import ugettext_lazy as _lazy
 
 import amo
-from amo.helpers import absolutify
 from amo.utils import JSONEncoder, to_language
 from mkt.access import acl
 from mkt.comm.utils import create_comm_note
@@ -21,7 +20,7 @@ from mkt.constants import comm
 from mkt.constants.features import FeatureProfile
 from mkt.files.models import File
 from mkt.reviewers.models import EscalationQueue, RereviewQueue, ReviewerScore
-from mkt.site.helpers import product_as_dict
+from mkt.site.helpers import absolutify, product_as_dict
 from mkt.site.mail import send_mail_jinja
 from mkt.webapps.models import Webapp
 from mkt.webapps.tasks import set_storefront_data
