+++ mkt/webapps/management/commands/convert_icons.py
@@ -7,8 +7,8 @@ from django.core.files.storage import default_storage as storage
 
 import amo
 from amo.decorators import write
-from amo.storage_utils import walk_storage
 from amo.utils import chunked, resize_image
+from mkt.site.storage_utils import walk_storage
 from mkt.webapps.models import Addon
