+++ mkt/webpay/views.py
@@ -16,7 +16,7 @@ from rest_framework.response import Response
 from rest_framework.viewsets import GenericViewSet
 
 import amo
-from amo.helpers import absolutify, urlparams
+from amo.helpers import urlparams
 from lib.cef_loggers import app_pay_cef
 from mkt.api.authentication import (RestAnonymousAuthentication,
                                     RestOAuthAuthentication,
@@ -26,6 +26,7 @@ from mkt.api.base import CORSMixin, MarketplaceView
 from mkt.purchase.models import Contribution
 from mkt.receipts.utils import create_inapp_receipt
 from mkt.site.mail import send_mail_jinja
+from mkt.site.helpers import absolutify
 from mkt.webpay.forms import FailureForm, PrepareInAppForm, PrepareWebAppForm
 from mkt.webpay.models import ProductIcon
 from mkt.webpay.serializers import ProductIconSerializer
