--- mkt/zadmin/templates/zadmin/addon_manage.html
@@ -1,72 +1,0 @@
-{% extends 'zadmin/base.html' %}
-
-{% block title %}{{ mkt_page_title('Manage {0}')|fe(addon.name) }}{% endblock %}
-
-{% block content %}
-<h2>Manage {{ addon.name }}</h2>
-<form method="post" action="">
-  <p>
-    <a href="{{ addon.get_detail_url() }}">Public Listing</a> |
-    <a href="{{ addon.get_dev_url() }}">Edit App</a>
-  </p>
-  <p>
-    <label>Status:</label> {{ form.status }}<br>
-    <label>Highest:</label> {{ form.highest_status }}<br>
-    <label>In queues:</label>
-    {% if addon.in_escalation_queue() %}
-      <a href="{{ url('reviewers.apps.review', addon.app_slug) }}">escalation</a>
-    {% endif %}
-    {% if addon.in_rereview_queue() %}
-      <a href="{{ url('reviewers.apps.review', addon.app_slug) }} ">re-review</a>
-    {% endif %}
-    <br/>
-  </p>
-  {% if versions %}
-  <h3>Versions &amp; Files</h3>
-    <table>
-      <thead>
-        <tr>
-          <th>Date</th>
-          <th>Version ID</th>
-          <th>File ID</th>
-          <th>Status</th>
-          <th>Hash</th>
-        </tr>
-      </thead>
-      <tbody>
-      {% for v in versions %}
-        <tr>
-          <td>{{ v.created|datetime }}</td>
-          <td><a href="{{ url('reviewers.apps.review', addon.app_slug) }}" title="{{ v.version }}">{{ v.id }}</a></td>
-          {% with files = file_map.get(v.id, []) %}
-            {% if files %}
-              {% for file in files %}
-                {% if not loop.first %}</tr><tr><td colspan="2">{% endif %}
-                <td>{{ file.id }} - {{ file.filename }}</td>
-                <td>
-                  {{ form_map[file.id].status }}
-                  {% for hidden in form_map[file.id].hidden_fields() %}
-                    {{ hidden }}
-                  {% endfor %}
-                </td>
-                <td><span title="{{ file.hash }}">{{ file.hash[:13] }}...{{ file.hash[-6:] }}</td>
-              {% endfor %}
-            {% else %}
-              <td colspan="4"></td>
-            {% endif %}
-          {% endwith %}
-        </tr>
-      {% endfor %}
-    </table>
-  {% endif %}
-  <p><input type="submit" value="Update Statuses"></p>
-  {{ formset.management_form }}
-  {{ csrf() }}
-  {% if pager.paginator.num_pages > 1 %}
-    <div class="listing-footer">
-      {{ pager|paginator }}
-    </div>
-  {% endif %}
-</form>
-
-{% endblock %}
